+++ CHANGELOG.md
@@ -189,6 +189,7 @@ here is a non-exhaustive list of significant departures from the original gem:
 * Explicitly document modifying item tags, labels and categories (where possible), and notify openHAB of the change
 * {OpenHAB::Core::Events::ItemStateEvent} and {OpenHAB::Core::Events::ItemStateChangedEvent} now have full sets of predicate methods.
 * {OpenHAB::DSL::Rules::Terse terse rules} now have an `on_load` parameter.
+* {Item#all_groups Item#all_groups}, {Enumerable#all_members}, {Enumerable#groups}, {Enumerable#all_groups}.
 
 ### Bug Fixes